+++ slony/dlgRepCluster.cpp
@@ -891,7 +891,7 @@ void dlgRepCluster::OnOK(wxCommandEvent &ev)
 			done = remoteConn->ExecuteVoid(
 			           wxT("SELECT ") + schemaPrefix + wxT("storepath(") +
 			           txtNodeID->GetValue() + wxT(", ") +
-			           NumToStr((long)cbAdminNode->wxItemContainer::GetClientData(cbAdminNode->GetCurrentSelection())) + wxT(", ") +
+			           NumToStr((long long)cbAdminNode->wxItemContainer::GetClientData(cbAdminNode->GetCurrentSelection())) + wxT(", ") +
 			           qtDbString(wxT("host=") + database->GetServer()->GetName() +
 			                      wxT(" port=") + NumToStr((long)database->GetServer()->GetPort()) +
 			                      wxT(" dbname=") + database->GetName()) + wxT(", ")
@@ -930,7 +930,7 @@ void dlgRepCluster::CheckChange()
 	if (cluster)
 	{
 		int sel = cbAdminNode->GetCurrentSelection();
-		bool changed = (sel >= 0 && (long)cbAdminNode->wxEvtHandler::GetClientData() != cluster->GetAdminNodeID());
+		bool changed = (sel >= 0 && (long long)cbAdminNode->wxEvtHandler::GetClientData() != cluster->GetAdminNodeID());
 
 		EnableOK(changed || txtComment->GetValue() != cluster->GetComment());
 	}
@@ -1039,7 +1039,7 @@ wxString dlgRepCluster::GetSql()
 		int sel = cbAdminNode->GetCurrentSelection();
 		if (sel >= 0)
 		{
-			long id = (long)cbAdminNode->wxItemContainer::GetClientData(sel);
+			long id = (long long)cbAdminNode->wxItemContainer::GetClientData(sel);
 			if (id != cluster->GetAdminNodeID())
 				settings->WriteLong(wxT("Replication/") + cluster->GetName() + wxT("/AdminNode"), id);
 		}
